Enhances request function to handle https
Renamed original request function to request_http, added function
request_https with the same signature and functionality, but that uses
urllib.HTTPSConnection to communicate over https.
Added new function called request, that has the same signature as the
original function request but that is able to handle http as well as
https urls.